+++ server/src/main/java/org/opensearch/index/translog/WriteOnlyTranslogManager.java
@@ -15,6 +15,7 @@ import org.opensearch.index.shard.ShardId;
 import org.opensearch.index.translog.listener.TranslogEventListener;
 
 import java.io.IOException;
+import java.util.function.BooleanSupplier;
 import java.util.function.LongSupplier;
 import java.util.function.Supplier;
 
@@ -36,7 +37,8 @@ public class WriteOnlyTranslogManager extends InternalTranslogManager {
         String translogUUID,
         TranslogEventListener translogEventListener,
         LifecycleAware engineLifecycleAware,
-        TranslogFactory translogFactory
+        TranslogFactory translogFactory,
+        BooleanSupplier primaryModeSupplier
     ) throws IOException {
         super(
             translogConfig,
@@ -49,7 +51,8 @@ public class WriteOnlyTranslogManager extends InternalTranslogManager {
             translogUUID,
             translogEventListener,
             engineLifecycleAware,
-            translogFactory
+            translogFactory,
+            primaryModeSupplier
         );
     }